RE-1526 Browsable Artifacts Attempt 2
The first attempt at this failed, as we were using a cloud files
api call which takes an archive and expands it out into a container.
That operation was too slow and unreliable to be useful.

As a recap, the aim is to have artifacts be browsable (and therefore
linkable) but also downloadable in bulk for local analysis.

This commit achieves both aims, by uploading a single archive but
also uploading each file individually. Both of these operations
have potential to take significant amounts of time. Archive creation
has taken 3 hours in the past due to compression. In order to speed
that up, this commit switches from bzip compression to gzip --fast
which can be an order of magnitude faster. Single file upload can
also be slow due to the number of api operations it requires. The
api rate limit is 100r/s so the upload is done with 100 threads.

In my testing, compression and both upload operations for an RPCO
artifact set took 7 minutes. That included 20k files totalling ~450mb.